requirements.txt
alembic==1.5.4
attrs==20.3.0
backcall==0.2.0
click==7.1.2
coverage==5.4
decorator==4.4.2
factory-boy==2.12.0
Faker==6.1.1
flake8==3.8.4
Flask==1.1.2
flask-marshmallow==0.11.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
inflection==0.5.1
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
mccabe==0.6.1
more-itertools==8.7.0
packaging==20.9
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.8.0
pyparsing==2.4.7
pytest==5.4.1
pytest-cov==2.8.1
pytest-factoryboy==2.0.3
pytest-flask==1.0.0
python-dateutil==2.8.1
python-decouple==3.3
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.23
text-unidecode==1.3
traitlets==5.0.5
wcwidth==0.2.5
webargs==5.5.3
Werkzeug==1.0.1
